+++ src/main/java/com/dayaeyak/waiting/domain/repository/jpa/WaitingRepository.java
@@ -6,14 +6,14 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
+import java.lang.Long;
 import java.util.List;
 
 @Repository
-public interface WaitingRepository extends JpaRepository<Waiting, BigInteger> {
+public interface WaitingRepository extends JpaRepository<Waiting, Long> {
     Waiting save(Waiting waiting);
-    Page<Waiting> findByRestaurantIdAndDeletedAtIsNull(BigInteger restaurantId, Pageable pageable);
-    List<Waiting> findByRestaurantIdAndDeletedAtIsNull(BigInteger restaurantId);
+    Page<Waiting> findByRestaurantIdAndDeletedAtIsNull(Long restaurantId, Pageable pageable);
+    List<Waiting> findByRestaurantIdAndDeletedAtIsNull(Long restaurantId);
 
 }
